09_testing/tdd.py
'''
Write a script that demonstrates TDD. Using pseudocode, plan out a couple simple functions. They could be
as simple as add and subtract or more complex such as functions that read and write to files.

Instead of writing out the functions, only provide the tests. Think about how the functions might
fail and write tests that will check and prevent failure.

You do not need to implement the actual functions after writing the tests but you may.

'''


# creating functions with input parameters
def add(x, y):
    return x + y


def sub(x, y):
    return x - y


def multi(x, y):
    return x * y


def divide(x, y):
    try:
        return x/y
    except ZeroDivisionError as zde:
        print("You cannot divide by zero! Please enter a positive number.")
        return zde
